requirements.txt
Flask==2.0.1
Pillow==8.3.1
numpy==1.21.1
opencv-python-headless==4.5.3.56
ezdxf==0.16.5
gunicorn==20.1.0
flask-cors==3.0.10 